add second handler
